+++ src/BabylonStudio/rtc/log_system.h
