+++ src/BabylonStudio/rtc/playground_autocompile.h
@@ -1,8 +1,8 @@
-#include <SamplesRunner/rtc/suppress_warnings.h>
-#include <SamplesRunner/rtc/link_libraries.h>
+#include <BabylonStudio/rtc/suppress_warnings.h>
+#include <BabylonStudio/rtc/link_libraries.h>
 #include <RuntimeObjectSystem/ObjectInterfacePerModule.h>
-#include <SamplesRunner/rtc/iscene_producer.h>
-#include <SamplesRunner/rtc/interface_ids.h>
+#include <BabylonStudio/rtc/iscene_producer.h>
+#include <BabylonStudio/rtc/interface_ids.h>
 
 class Playground : public TInterface<BABYLON::rtc::IID_RENDERABLESCENE_PRODUCER, BABYLON::rtc::ISceneProducer>
 {
