+++ src/BabylonStudio/rtc/suppress_warnings.h
