+++ src/BabylonStudio/spawn_screenshots.cpp
